rename Log class to Augment
